requirements.txt
tensorflow
redis
pandas
gym
grpcio
grpcio-tools